+++ unnamed_engine/src/event/event.rs
@@ -8,11 +8,17 @@
 //                 ╚═════╝ ╚═╝  ╚═══╝╚══════╝╚═╝  ╚═══╝ LIB
 //
 ////////////////////////////////////////////////////////////////////////////////
-// ? Defines the events that will be passed to the applications.
+// ? Defines the events that will be used internally and passed down to the
+// ? applications.
 
 pub use winit::keyboard::KeyCode;
 
 pub enum Event {
+    Shutdown,
+    Resize {
+        width: u32,
+        height: u32,
+    },
     Keyboard {
         key: KeyCode,
         is_pressed: bool,